pyproject.toml
[tool.poetry]
name = "tasktick"
version = "0.0.5"
description = ""
authors = ["dnlzrgz <24715931+dnlzrgz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-ninja = "^1.1.0"
environs = "^11.0.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
pre-commit = "^3.6.2"
commitizen = "^3.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
